--- cart-service/src/main/java/com/example/CartService/CartServiceApplication.java
@@ -1,14 +1,0 @@
-package com.example.CartService;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients
-public class CartServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CartServiceApplication.class, args);
-    }
-}
