--- cart-service/src/main/java/com/example/CartService/DTO/CartItemDTO.java
@@ -1,11 +1,0 @@
-package com.example.CartService.DTO;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class CartItemDTO {
-    private String productId;
-    private LocalDateTime addedAt = LocalDateTime.now();
-}
