--- cart-service/src/main/java/com/example/CartService/exception/BadRequestException.java
@@ -1,8 +1,0 @@
-package com.example.CartService.exception;
-
-public class BadRequestException extends RuntimeException {
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
-
