--- cart-service/src/main/java/com/example/CartService/exception/ErrorDetail.java
@@ -1,9 +1,0 @@
-package com.example.CartService.exception;
-
-import lombok.Data;
-
-@Data
-public class ErrorDetail {
-    private String code;
-    private String message;
-}
