--- cart-service/src/main/java/com/example/CartService/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.CartService.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
