--- cart-service/src/main/java/com/example/CartService/model/Cart.java
@@ -1,21 +1,0 @@
-package com.example.CartService.model;
-
-import lombok.Data;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@Document(
-        "carts"
-)
-public class Cart {
-    @Id
-    private String id;
-    private int totalPrice;
-    private List<CartItem> cartItems = new ArrayList<>();
-    private int totalItem;
-    private String userId;
-}
