--- cart-service/src/main/java/com/example/CartService/model/CartItem.java
@@ -1,12 +1,0 @@
-package com.example.CartService.model;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class CartItem {
-    private String productId;
-    private int quantity = 1;
-    private LocalDateTime addedAt;
-}
