+++ cart-service/src/main/java/com/example/cart_service/CartServiceApplication.java
@@ -1,0 +1,13 @@
+package com.example.cart_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CartServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CartServiceApplication.class, args);
+	}
+
+}
