+++ cart-service/src/main/java/com/example/cart_service/DTO/CartResponse.java
@@ -1,0 +1,17 @@
+package com.example.cart_service.DTO;
+
+import lombok.Data;
+import org.springframework.data.annotation.Id;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Data
+public class CartResponse {
+    @Id
+    private String id;
+    private int totalPrice;
+    private List<ProductDTO> products = new ArrayList<>();
+    private int totalItem;
+    private String userId;
+}
