+++ cart-service/src/main/java/com/example/cart_service/ENUM/PRODUCT_STATE.java
@@ -1,0 +1,6 @@
+package com.example.cart_service.ENUM;
+
+public enum PRODUCT_STATE {
+    ACTIVE,
+    HIDDEN
+}
