+++ cart-service/src/main/java/com/example/cart_service/client/ProductClient.java
@@ -1,0 +1,20 @@
+package com.example.cart_service.client;
+
+import com.example.cart_service.DTO.ProductDTO;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
+import java.util.List;
+
+@FeignClient(name = "product-service", url = "http://localhost:4444/api/product")
+public interface ProductClient {
+
+    @GetMapping("/{id}")
+    ProductDTO getProductById(@PathVariable("id") String id);
+
+    @PostMapping("/bulk")
+    List<ProductDTO> getProductsByIds(@RequestBody List<String> ids);
+}
