+++ cart-service/src/main/java/com/example/cart_service/exception/BadRequestException.java
@@ -1,0 +1,8 @@
+package com.example.cart_service.exception;
+
+public class BadRequestException extends RuntimeException {
+    public BadRequestException(String message) {
+        super(message);
+    }
+}
+
