+++ cart-service/src/main/java/com/example/cart_service/exception/ErrorDetail.java
@@ -1,0 +1,9 @@
+package com.example.cart_service.exception;
+
+import lombok.Data;
+
+@Data
+public class ErrorDetail {
+    private String code;
+    private String message;
+}
