+++ cart-service/src/main/java/com/example/cart_service/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.cart_service.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
