+++ cart-service/src/main/java/com/example/cart_service/model/CartItem.java
@@ -1,0 +1,21 @@
+package com.example.cart_service.model;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class CartItem {
+    private String productId;
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    private int quantity = 1;
+    private LocalDateTime addedAt;
+}
