+++ cart-service/src/main/resources/application.properties
@@ -1,10 +1,2 @@
-spring.application.name=CartService
-spring.datasource.url=jdbc:mysql://localhost:3306/cartdb
-spring.datasource.username=root
-spring.datasource.password=ArandomPassword5@
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.show-sql=true
-spring.data.redis.host=localhost
-spring.data.redis.port=6379
-spring.data.mongodb.uri=mongodb://localhost:27017/your_database_name
+spring.application.name=cart-service
+spring.config.import=optional:configserver:http://localhost:8888
