--- cart-service/src/test/java/com/example/CartService/CartServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.CartService;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CartServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
