+++ cart-service/src/test/java/com/example/cart_service/CartServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.cart_service;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CartServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
